requirements.txt
asgiref==3.2.3
astroid==2.3.3
bcrypt==3.1.7
cffi==1.13.2
cryptography==2.8
Django==3.0.2
django-environ==0.4.5
dnspython==1.16.0
ecdsa==0.15
fabric==2.5.0
gunicorn==20.0.4
invoke==1.4.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysqlclient==1.4.6
paramiko==2.7.1
protobuf==3.6.1
pycparser==2.19
pycrypto==2.6.1
pylint==2.4.4
PyNaCl==1.3.0
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
typed-ast==1.4.1
wrapt==1.11.2
